requirements.txt
fastapi
pydantic
uvicorn
python-decouple
peewee
PyMySQL